# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Instantly.ai clone API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /labels:
    get:
      operationId: listLabels
      tags: [labels]
      summary: List all labels
      responses:
        "200":
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
    post:
      operationId: createLabel
      tags: [labels]
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelInput"
      responses:
        "201":
          description: Created label
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
  /labels/{id}:
    delete:
      operationId: deleteLabel
      tags: [labels]
      summary: Delete a label
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Label:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
    LabelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
...